+++ contest/AtCoder/arc049/A.cpp
@@ -2,11 +2,11 @@
 
 int main() {
   string s;
-  vector<int> a(4);
+  vector<uint> a(4);
   cin >> s >> a;
-  for (int i = 0; i <= (int)s.size(); ++i) {
+  for (uint i = 0; i <= s.size(); ++i) {
     if (in(a, i)) cout << "\"";
-    if (i < (int)s.size()) cout << s[i];
+    if (i < s.size()) cout << s[i];
   }
   cout << endl;
 }
