+++ contest/yukicoder/179.cpp
@@ -1,12 +1,12 @@
 #include "vector.hpp"
 
 bool solve(vector<string> s, int y, int x) {
-  for (int i = 0; i < (int)s.size(); ++i) {
-    for (int j = 0; j < (int)s[0].size(); ++j) {
+  for (uint i = 0; i < s.size(); ++i) {
+    for (uint j = 0; j < s[0].size(); ++j) {
       if (s[i][j] == '#') {
         s[i][j] = '.';
-        if (i + y >= (int)s.size()) return false;
-        if (j + x >= (int)s[0].size()) return false;
+        if (i + y >= s.size()) return false;
+        if (j + x >= s[0].size()) return false;
         if (s[i + y][j + x] == '.') return false;
         s[i + y][j + x] = '.';
       }
@@ -16,7 +16,7 @@ bool solve(vector<string> s, int y, int x) {
 }
 
 bool solve(vector<string>& s) {
-  for (int i = 0; i < (int)s.size(); ++i) for (int j = 0; j < (int)s[0].size(); ++j) if (solve(s, i, j)) return true;
+  for (uint i = 0; i < s.size(); ++i) for (uint j = 0; j < s[0].size(); ++j) if (solve(s, i, j)) return true;
   return false;
 }
