+++ include/geometry/convex_polygon.hpp
@@ -12,11 +12,11 @@ public:
   ConvexPolygon(vector<Point> points) {
     int flag = ~(strict ? LEFT : LEFT | FRONT);
     sort(points.begin(), points.end());
-    for (int i = 0; i < (int)points.size(); emplace_back(points[i++])) {
+    for (uint i = 0; i < points.size(); emplace_back(points[i++])) {
       while (size() > 1u && ccw(*(end() - 2), back(), points[i]) & flag) pop_back();
     }
     for (int i = points.size() - 2, r = size(); i >= 0; emplace_back(points[i--])) {
-      while ((int)size() > r && ccw(*(end() - 2), back(), points[i]) & flag) pop_back();
+      while (int(size()) > r && ccw(*(end() - 2), back(), points[i]) & flag) pop_back();
     }
     pop_back();
   }
