+++ include/graph/clique.hpp
@@ -39,7 +39,7 @@ public:
   void maximum(vector<int>& maximumClique, vector<int>& current, vector<int> remains) {
     if (maximumClique.size() < current.size()) maximumClique = current;
     int maxSize = maximumClique.size();
-    if ((int)current.size() + (int)remains.size() <= maxSize) return;
+    if (int(current.size() + remains.size()) <= maxSize) return;
     if (graph.getDegree(remains[maxSize - current.size()]) < maxSize) return;
     for (int v : remains) {
       current.emplace_back(v);
